added check to make sure a user can't delete a blog that they did not make.

+++ src/server/database/queries/Blogs.ts
@@ -27,7 +27,7 @@ Where Blogs.id = ?;
     [id]
   );
 
-const destroyOneBlog = (id: number) => query("DELETE FROM Blogs WHERE id = ?", [id]);
+const destroyOneBlog = (id: number, authorid: number) => query("DELETE FROM Blogs WHERE id = ? AND authorid = ?", [id, authorid]);
 const destroyOneBlogTag = (id: number) => query("DELETE FROM BlogTags WHERE blogid = ?", [id]);
 const createOneBlog = (newBlog: NewBlog) => query("INSERT INTO Blogs SET ?", [newBlog]);
 const updateOneBlog = (updatedBlog: NewBlog, id: number) => query("UPDATE Blogs SET ? WHERE id=?", [updatedBlog, id]);
